Add files via upload
The updated code is available in the below GIT path after forking from the given baseline –
https://github.com/Vasanth2796/HPC_HW5_RL

The codes which needs to be considered for evaluation are –
1. sb3_highway_dqn. ipynb
2. sb3_highway_ppo. Ipynb

Both these files are included in the uploaded ZIP folder (AuE8930_HPC_HW5_RL_VasanthSeethapathi.zip) including the visualization graphs for both training and testing. Also, I have added the comments for “sb3_highway_ppo. Ipynb” file.
This ZIP folder is uploaded in the above mentioned GIT path.